+++ src/extension.ts
@@ -2,21 +2,34 @@
 
 import * as vscode from "vscode";
 import Linter from "./Linter";
+import { debounce } from 'ts-debounce';
 
 export function activate(context: vscode.ExtensionContext) {
   const linter = new Linter();
   context.subscriptions.push(linter);
 
-  const updateDiagnostics = (document: vscode.TextDocument) => {
+  const updateDiagnostics = (doc: vscode.TextDocument|vscode.TextDocumentChangeEvent) => {
+    let document: vscode.TextDocument;
+    if (doc.hasOwnProperty('document')) {
+      document = (<vscode.TextDocumentChangeEvent>doc).document;
+    }
+    else{
+      document = <vscode.TextDocument>doc;
+    }
     linter.run(document);
   };
+  const debouncedFn = debounce(updateDiagnostics, 800);
 
   context.subscriptions.push(
-    vscode.workspace.onDidSaveTextDocument(updateDiagnostics)
+    vscode.workspace.onDidSaveTextDocument(debouncedFn)
   );
 
   context.subscriptions.push(
-    vscode.workspace.onDidOpenTextDocument(updateDiagnostics)
+    vscode.workspace.onDidOpenTextDocument(debouncedFn)
+  );
+
+  context.subscriptions.push(
+    vscode.workspace.onDidChangeTextDocument(debouncedFn)
   );
 
   context.subscriptions.push(
@@ -25,7 +38,7 @@ export function activate(context: vscode.ExtensionContext) {
     })
   );
 
-  vscode.workspace.textDocuments.forEach(updateDiagnostics);
+  vscode.workspace.textDocuments.forEach(debouncedFn);
 }
 
 export function deactivate() {}