+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.0.2
+
+- Linting on change, not just on save
+- Linting files without a working folder
+- Debounce lint operations
+
 ## 0.0.1
 
 - Initial release